requirements.txt
flask==2.0.1
scikit-learn==1.0.2
pandas==1.3.3
numpy==1.21.2
nltk==3.6.3
flask-wtf==0.15.1
flask-sqlalchemy==2.5.1
flask-login==0.5.0
gunicorn==20.1.0
python-dotenv==0.19.1
